nrf_modem_lib: fix trace data size return type
nrf_modem_lib_trace_data_size() could wrap when returning a negative errno
because it used an unsigned type. Switch to signed int so -ENOTSUP etc are
returned correctly.

size_t evaluates to unsigned int in our toolchain.

+++ include/modem/nrf_modem_lib_trace.h
@@ -73,7 +73,7 @@ int nrf_modem_lib_trace_level_set(enum nrf_modem_lib_trace_level trace_level);
  *
  * @returns Number of bytes stored in the trace backend.
  */
-size_t nrf_modem_lib_trace_data_size(void);
+int nrf_modem_lib_trace_data_size(void);
 
 /**
  * @brief Read trace data
